backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
slowapi==0.1.9
secure==0.3.0

# Caching
redis==5.0.1
cachetools==5.3.2

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx>=0.26.0

# Email (async)
aiosmtplib>=3.0.0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3

# PDF Generation
reportlab==4.0.9
Pillow==10.2.0

# Geometry
shapely

# Database & Authentication
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
bcrypt>=4.0.0
python-multipart>=0.0.9
email-validator>=2.0.0

# AI Models (Module Communication)
google-generativeai
groq
anthropic>=0.40.0
json-repair>=0.28.0

# Excel Manipulation (Module Commerce)
pandas>=2.0.0
openpyxl>=3.1.0
